🏗️ Professional project restructure with GUI executable
✨ COMPLETE FRONTEND/BACKEND SEPARATION:
📁 Frontend (GUI/Interface):
├── frontend/modern_gui.py    # 🎨 Modern GUI (main interface)
├── frontend/gui.py           # 🖥️ Classic GUI
├── frontend/cloud_sync.py    # ☁️ Cloud integration interface
└── frontend/__init__.py      # Package initialization

📁 Backend (Core Logic):
├── backend/                  # Core CLI functionality
├── main.py                   # 💻 CLI entry point
└── backend/__init__.py       # Package initialization

🚀 NEW LAUNCHER SYSTEM:
├── launcher.py               # 🎯 Interface selector menu
└── build.py                  # 🔨 Enhanced build script

📦 EXECUTABLE INCLUDED (10MB):
├── release/PyTo-Do-GUI.exe   # Ready-to-run GUI application
└── release/README.md         # Usage instructions

🎯 ENHANCED FEATURES:
- Professional project structure
- Multiple interface options via launcher
- Proper Python package organization
- GUI executable included in repository
- Updated build system with more options
- Proper attribution to CosmicLM throughout

👥 PERFECT FOR USERS:
- python launcher.py - Choose any interface
- python frontend/modern_gui.py - Direct GUI launch
- ./release/PyTo-Do-GUI.exe - No Python needed!
