+++ frontend/__init__.py
@@ -1,0 +1,8 @@
+"""
+PyTo-Do Frontend Package
+Contains all GUI and interface components
+Originally created by CosmicLM | Enhanced by mdnoyon9758
+"""
+
+__version__ = "1.1.0"
+__author__ = "CosmicLM (original) | mdnoyon9758 (enhanced)"
